requirements.txt
nasdaq-datalink>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
pandas-datareader>=0.10.0
fredapi>=0.5.0
pyyaml>=6.0.0
requests>=2.31.0
